+++ node/src/syncer.rs
@@ -26,7 +26,7 @@ use tracing::{debug, info, info_span, instrument, warn, Instrument};
 
 use crate::executor::{sleep, spawn, spawn_cancellable, Interval};
 use crate::p2p::{P2p, P2pError};
-use crate::store::{calculate_missing_ranges, HeaderRanges, Store, StoreError};
+use crate::store::{calculate_missing_ranges, utils::HeaderRanges, Store, StoreError};
 use crate::utils::OneshotSenderExt;
 
 type Result<T, E = SyncerError> = std::result::Result<T, E>;
